+++ TombEngine/Objects/Effects/EmberEmitter.h
