+++ TombEngine/Objects/TR5/tr5_objects.cpp
@@ -55,7 +55,7 @@
 #include "Objects/TR5/Object/tr5_twoblockplatform.h"
 
 // Traps
-#include "Objects/Effects/ember_emitter.h"
+#include "Objects/Effects/EmberEmitter.h"
 #include "Objects/Effects/tr5_electricity.h"
 #include "Objects/TR5/Trap/LaserBarrier.h"
 #include "Objects/TR5/Trap/ZipLine.h"
